waker: avoid double borrowing task
tasks may wake themselves, and if the executor enqueues the task
immediately it will fail to borrow the task - which has already been
borrowed to call .poll().

we can send these tasks into the wake pipe, which will pull them out
and enqueue them later - when we know no tasks are borrowed.